requirements.txt
streamlit==1.32.0
pandas>=2.2.0
altair==5.2.0
sqlalchemy==2.0.35
python-dotenv==1.0.1
